fix(material-experimental/mdc-paginator): rename module file
