+++ src/material-experimental/mdc-paginator/module.ts
