+++ src/material-experimental/mdc-paginator/public-api.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export * from './paginator-module';
+export * from './module';
 export * from './paginator';
 export {
   MatPaginatorIntl,
